backend/requirements.txt
Django
djangorestframework
psycopg2-binary
django-cors-headers
python-dotenv
langchain
langgraph
langchain-google-genai